src/main/java/web/lose/LoseSecondServlet.java
package web.lose;

import models.AnswerChallenge;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import service.SecondService;

import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

@WebServlet("/losesecond")
public class LoseSecondServlet extends HttpServlet {
    private static final Logger LOGGER = LoggerFactory.getLogger(LoseSecondServlet.class);
    SecondService secondService = new SecondService();
    @Override
    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        LOGGER.info("Method doPost() LoseSecondServlet started");
        AnswerChallenge answer = secondService.call(Boolean.parseBoolean(req.getParameter("answer")));

        resp.setStatus(201);
        req.setAttribute("answer", answer.getMessage());
        req.getRequestDispatcher(answer.getPage()).forward(req, resp);
        LOGGER.info("Method doPost() LoseSecondServlet finished");
    }
}
